split opto into light and opsins
